+++ sheets/PointOfInterestSheet.js
@@ -11,7 +11,7 @@ export class PointOfInterestSheet extends EnhancedJournalSheet {
             title: i18n("MonksEnhancedJournal.PointOfInterest"),
             template: "modules/monks-enhanced-journal/templates/poi.html",
             tabs: [{ navSelector: ".tabs", contentSelector: ".sheet-body", initial: "description" }],
-            scrollY: [".description"]
+            scrollY: [".tab.description .tab-inner"]
         });
     }